% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{anti_t}
\alias{anti_t}
\title{Anti transpose (internal)}
\usage{
anti_t(m)
}
\arguments{
\item{m}{square matrix}
}
\value{
the anti traspose of \code{m}
}
\description{
Anti transpose (internal)
}
\keyword{internal}
