% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{B0}
\alias{B0}
\title{Generate a naive stable matrix}
\usage{
B0(p)
}
\arguments{
\item{p}{dimension of the matrix}
}
\value{
a stable matrix with off-diagonal entries equal to 1 and
diagonal entries equal to \code{-p}
}
\description{
Generate a naive stable matrix
}
