% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{deltaGraph}
\alias{deltaGraph}
\title{Difference graph}
\usage{
deltaGraph(B1, B2, legend = FALSE, ...)
}
\arguments{
\item{B1}{matrix}

\item{B2}{matrix}

\item{legend}{logical}

\item{...}{parameters to be passed to the igraph plotting}
}
\value{
a graph with colored edges where
in red are plotted the edges in \code{B1} and not in
\code{B2} and in blue the edges in \code{B2} and not in
\code{B1}.
}
\description{
Difference graph
}
