% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclm.R
\name{gclm}
\alias{gclm}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Generalized Cross-Lagged Panel Models (Zyphur et al., 2019)}
\usage{
gclm(xvar, yvar, waves)
}
\arguments{
\item{xvar}{item stem (and delimeter, if present) for x variables}

\item{yvar}{item stem (and delimeter, if present) for y variables}

\item{waves}{number of waves to model (i.e., number following each item stem)}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions, while also exporting to current working directory
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying dyadic configural, loading, and intercept invariant
measurement models for either a specified X or Y factor.
}
\examples{
base.model = gclm(xvar = "sacrifice", yvar = "satisfaction", waves = 5)
}
