% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{genGlasso}
\alias{genGlasso}
\title{generalized graphical lasso}
\usage{
genGlasso(Sigma, P = diag(nrow(Sigma)), G = -diag(nrow(Sigma)),
  lambda = 0.1, maxIter = 1000, eps = 1e-05, alpha = 0.5,
  beta = 0.5, trace = 0)
}
\arguments{
\item{Sigma}{the covariance matriz}

\item{P}{initial precision matrix}

\item{G}{Matrix in the generlaized penalization}

\item{lambda}{penalization coefficient}

\item{maxIter}{maximum number of iterations}

\item{eps}{threshold for stopping criteria}

\item{alpha}{param for line search}

\item{beta}{param for line search}

\item{trace}{integer, if >0 print info}
}
\value{
The estimated precision matrix
}
\description{
Solve the generalized graphical lasso problem with proximal gradient
}
