% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hamming}
\alias{hamming}
\title{Hamming distance between adjacency matrices}
\usage{
hamming(A, B)
}
\arguments{
\item{A}{matrix}

\item{B}{matrix}
}
\value{
The hamming distance between the directed graphs induced by
\code{A} and \code{B}.
}
\description{
Hamming distance between adjacency matrices
}
