% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hessianll}
\alias{hessianll}
\alias{diagHessianll}
\title{Hessian of log-likelihood for Lyapunov parametrization}
\usage{
hessianll(B, C = diag(nrow(B)), Sigma)

diagHessianll(B, C = diag(nrow(B)), Sigma)
}
\arguments{
\item{B}{the B matrix}

\item{C}{the C matrix}

\item{Sigma}{the empirical covariance matrix}
}
\value{
the Hessian matrix of the log-likelihood
}
\description{
Hessian of log-likelihood for Lyapunov parametrization
}
