% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{listInverseBlocks}
\alias{listInverseBlocks}
\title{List all the inverse matrices of the leading sub-matrix of P=Sigma^{-1}
(INTERNAL)}
\usage{
listInverseBlocks(Sigma)
}
\arguments{
\item{Sigma}{an invertible  matrix}
}
\value{
a list with the inverse of the leading sub-matrices of Sigma^(-1)
}
\description{
List all the inverse matrices of the leading sub-matrix of P=Sigma^{-1}
(INTERNAL)
}
\keyword{internal}
