% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{llBpath}
\alias{llBpath}
\title{Path of B estimates (log-lik)}
\usage{
llBpath(Sigma, lambdas = NULL, C = diag(nrow(Sigma)), B0 = NULL,
  eps = 1e-08, maxIter = 1000, job = 11)
}
\arguments{
\item{Sigma}{empirical covariance matrix}

\item{lambdas}{increasing sequence of lambdas}

\item{C}{the known C = DD'}

\item{B0}{initial B matrix}

\item{eps}{convergence criteria}

\item{maxIter}{maximum iterations for each proximal gradient}

\item{job}{integer flag (0,1,10,11) as in \code{\link{proxgradllB}}}
}
\description{
Path of B estimates (log-lik)
}
