% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{llselectB}
\alias{llselectB}
\title{Select model in a path}
\usage{
llselectB(x, Sigma)
}
\arguments{
\item{x}{the path of solution, usually the output of \code{\link{llBpath}}}

\item{Sigma}{a covariance matrix}
}
\value{
the index of the selected model
}
\description{
Select model in a path
}
