% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mll}
\alias{mll}
\title{Minus Log-Likelihood for Gaussian model}
\usage{
mll(P, S)
}
\arguments{
\item{P}{the inverse of the covariance matrix}

\item{S}{the empirical covariance}
}
\description{
Minus Log-Likelihood for Gaussian model
}
