% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mllB}
\alias{mllB}
\title{Minus log-likelihood for the invariant distribution of OU}
\usage{
mllB(B, S, C = diag(nrow(B)))
}
\arguments{
\item{B}{coefficent matrix}

\item{S}{covariance matrix}

\item{C}{noise matrix}
}
\description{
Minus log-likelihood for the invariant distribution of OU
}
