% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clggm.R
\name{pnllbc}
\alias{pnllbc}
\title{Penalized-likelihood estimation of CLGGM}
\usage{
pnllbc(
  Sigma,
  B,
  C = diag(ncol(Sigma)),
  C0 = diag(ncol(Sigma)),
  eps = 0.01,
  alpha = 0.5,
  beta = 0.2,
  maxIter = 100,
  intitr = 100,
  lambda = 0,
  lambdac = 0,
  job = 0
)
}
\arguments{
\item{Sigma}{empirical covariance matrix}

\item{B}{initial B matrix}

\item{C}{intial C matrix}

\item{C0}{penalization matrix}

\item{eps}{convergence tolerance}

\item{alpha}{parameter line search}

\item{beta}{parameter line search}

\item{maxIter}{maximum number of iterations}

\item{intitr}{number of internal iterations}

\item{lambda}{penalization coefficient for B}

\item{lambdac}{penalization coefficient for C}

\item{job}{integer 0,1,10 or 11}
}
\value{
a list with the result of the optimization
}
\description{
Penalized-likelihood estimation of CLGGM
}
