% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clggm.R
\name{proxcdllB}
\alias{proxcdllB}
\title{Penalized likelihood estimation of CLGGM}
\usage{
proxcdllB(Sigma, B, C = diag(ncol(Sigma)), eps = 0.01, alpha = 0.5,
  maxIter = 1000, lambda = 0, job = 0)
}
\description{
Optimize the B matrix of a continuous Lyapunov
Gaussian graphical model (CLGGM) using proximal coordinate descent.
\deqn{\hat{B} = \arg \min_B LL(B,C) + \lambda||B||_1}
}
