% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rHurwitz}
\alias{rHurwitz}
\title{Random Stable (Hurwitz) matrix}
\usage{
rHurwitz(p = 1, d = 1)
}
\arguments{
\item{p}{integer, the dimension of the desired matrix}

\item{d}{fraction of non-zero entries}
}
\description{
Generate a random stable matrix, that is with all
eigenvalues with strict negative real part
}
