% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ROC}
\alias{ROC}
\alias{AUROC}
\alias{FPR}
\alias{TPR}
\title{Compute the ROC curve}
\usage{
ROC(x, y, n = 100)

AUROC(x, y, n = 100)

FPR(x, y)

TPR(x, y)
}
\arguments{
\item{x}{an array with confidence levels between 0 and 1}

\item{y}{the true target values (0 and >0)}

\item{n}{number of points}
}
\description{
Compute the ROC curve
}
