% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rOUinv}
\alias{rOUinv}
\title{Generate data from invariant distribution of O-U process}
\usage{
rOUinv(n = 1, B, D = diag(nrow = nrow(B), ncol = ncol(B)),
  mean = rep(0, nrow(B)))
}
\arguments{
\item{n}{number of sample}

\item{B}{square matrix, coefficients of the O-U equation}

\item{D}{noise matrix}

\item{mean}{invariant mean}
}
\value{
A list with components:
\itemize{
\item \code{sample} the produced sample
\item \code{Sigma} the covariance matrix of the invariant distribution
\item \code{C} the  C matrix obtained as DD'
\item \code{B} the coefficient square matrix
\item \code{mean} the mean vector of the invariant distribution
}
}
\description{
Sample observations from the invariant distribution of an
Ornstein-Uhlenbeck process:  \deqn{dX_t = B(X_t - \mu)dt + DdW_t}
}
