% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{RproxgradllB}
\alias{RproxgradllB}
\alias{gradllC}
\title{Estimate Continuous Lyapunov Gaussian GM}
\usage{
RproxgradllB(Sigma, B, C = diag(ncol(Sigma)), eps = 0.01,
  alpha = 0.2, beta = 0.5, maxIter = 1000, trace = 0, lambda = 0,
  r = FALSE, h = FALSE)

gradllC(Sigma, B, C = diag(ncol(Sigma)), C0 = diag(ncol(Sigma)),
  eps = 0.01, alpha = 0.2, beta = 0.5, maxIter = 1000, trace = 0,
  lambda = 0, t0 = 1)
}
\arguments{
\item{Sigma}{the empirical covariance matrix}

\item{B}{an initial guess for the coefficient matrix B}

\item{C}{the initial guess for the noise matrix C}

\item{eps}{stopping criteria}

\item{alpha}{parameter for line search}

\item{beta}{parameter for line search}

\item{maxIter}{maximum number of iterations}

\item{trace}{if >0 print info (1:termination message only,
2:messages at each step)}

\item{lambda}{penalization coefficient}

\item{r}{logical, if TRUE the set of non-zero parameters will be updated
at every iteration}

\item{h}{logical if TRUE only the non-zero entries of B will be updated}

\item{C0}{penalization matrix}

\item{t0}{initial step length}
}
\value{
the estimated B matrix (\code{estimateBLL}) or
the estiamted C matrix (\code{estiamteCLL}).
}
\description{
Estimate from cross sectional data
the coefficients and the noise term of an Ornstein-Uhlenbeck
process (\eqn{dX(t) = BX(t) + \sqrt{C} dW(t)})
using penalized maximum-likelihood.
}
