% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rSkewSymm}
\alias{rSkewSymm}
\title{Generate a random skew-symmetric matrix}
\usage{
rSkewSymm(p, rfun = rnorm)
}
\arguments{
\item{p}{integer the dimension of the desired matrix}

\item{rfun}{a function with signature \code{function(n)} that
generates \code{n} numbers (e.g. \code{rnorm})}
}
\value{
a skew-symmetric matrix
}
\description{
Generate a random skew-symmetric matrix
}
