% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rStableMetzler}
\alias{rStableMetzler}
\title{Generate stable Metzler matrix}
\usage{
rStableMetzler(p = 1, d = 1, lower = FALSE, rfun = rnorm,
  rdiag = rnorm)
}
\arguments{
\item{p}{integere the dimension of the matrix}

\item{d}{proportion of non zero entries}

\item{lower}{logical, should the matrix be lower triangular}

\item{rfun}{the random number generator for the matrix entries}

\item{rdiag}{the random number generator for the diagonal added error}
}
\value{
a square matrix with positive off-diagonal entries (Metzler)
and eigenvalues with strict negative real part (stable)
}
\description{
Generate stable Metzler matrix
}
