% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sylvester.R
\name{sylvester_c}
\alias{sylvester_c}
\alias{sylvester_cq}
\title{Link to fortran solver for Lyapunov equation}
\usage{
sylvester_c(A, E, QQ, cond.number = FALSE, all = FALSE)

sylvester_cq(A, E, QQ, WKV = rep(0, 2 * N^2 + 3 * N),
  cond.number = FALSE)
}
\description{
Link to fortran solver for Lyapunov equation
}
\keyword{internal}
