% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clggm.R
\name{clyap}
\alias{clyap}
\title{Solve continuous-time Lyapunov equations}
\usage{
clyap(B, C, Q = NULL, all = FALSE)
}
\arguments{
\item{B}{Square matrix}

\item{C}{Square matrix}

\item{Q}{Square matrix, the orthogonal matrix used
to transfrom the original equation}

\item{all}{logical}
}
\value{
The solution matrix \eqn{X}.
}
\description{
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{BX + XB' + C=0}
Using the Bartels-Stewart algorithm with Hessenberg–Schur
decomposition.
}
\details{
\code{clyap} uses lapack subroutines:
\code{DGEES} and \code{DTRSYL}
}
\examples{

}
