% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clggm.R
\name{clyap}
\alias{clyap}
\title{Solve continuous-time Lyapunov equations}
\usage{
clyap(B, C, Q = NULL, all = FALSE)
}
\arguments{
\item{B}{Square matrix}

\item{C}{Square matrix}

\item{Q}{Square matrix, the orthogonal matrix used
to transfrom the original equation}

\item{all}{logical}
}
\value{
The solution matrix \code{X} if \code{all = FALSE}. If
\code{all = TRUE} a list with components \code{X}, \code{B}
and \code{Q}. Where \code{B} and \code{Q} are the
Hessenberg-Schur form of the original matrix \code{B}
and the orthogonal matrix that performed the transformation.
}
\description{
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{BX + XB' + C=0}
Using the Bartels-Stewart algorithm with Hessenberg–Schur
decomposition. Optionally the Hessenberg-Schur
decomposition can be returned.
}
\details{
If the matrix \code{Q} is set then the matrix \code{B}
is assumed to be in upper quasi-triangular form
(Hessenberg-Schur canonical form),
as required by LAPACK subroutine \code{DTRSYL} and \code{Q} is
the orthogonal matrix associated with the Hessenberg-Schur form
of \code{B}.
Usually the matrix \code{Q} and the appropriate form of \code{B}
are obtained by a first call to \code{clyap(B, C, all = TRUE)}

\code{clyap} uses lapack subroutines:
\itemize{
\item \code{DGEES}
\item \code{DTRSYL}
}

Moreover, some additional subroutines are used from:
\itemize{
\item Algorithm 705; a FORTRAN-77 software package for
solving the Sylvester matrix equation AXBT + CXDT = E
}
}
\examples{

}
