% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clggm.R
\name{clyap}
\alias{clyap}
\alias{clyap2}
\alias{gradllB}
\alias{jacllB}
\title{Solve continuous-time Lyapunov/Sylvester equations
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{AXE' + EXA'+ Q=0.}}
\usage{
clyap(A, C, all = FALSE)

clyap2(A, Q, WKV, E)

gradllB(A, E, D, S, WKV, IX = NULL)

jacllB(A, E, S, WKV)
}
\arguments{
\item{A}{Square matrix}

\item{all}{logical}

\item{Q}{Square matrix}

\item{WKV}{the working vector obtained by a first call to lyapunov solver}

\item{E}{Square matrix or \code{NULL}}
}
\value{
The solution matrix \eqn{X}.
}
\description{
Solve continuous-time Lyapunov/Sylvester equations
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{AXE' + EXA'+ Q=0.}
}
\details{
\code{clyap} use \code{BKDIS}
from ALGORITHM 705, COLLECTED ALGORITHMS FROM ACM
}
\examples{

}
