% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclm.R
\name{gclm}
\alias{gclm}
\alias{gclm.path}
\title{Penalized-likelihood estimation of CLGGM}
\usage{
gclm(
  Sigma,
  B = -0.5 * diag(ncol(Sigma)),
  C = rep(1, ncol(Sigma)),
  C0 = rep(1, ncol(Sigma)),
  loss = "loglik",
  eps = 0.01,
  alpha = 0.5,
  maxIter = 100,
  lambda = 0,
  lambdac = 0,
  job = 0
)

gclm.path(
  Sigma,
  lambdas = NULL,
  B = -0.5 * diag(ncol(Sigma)),
  C = rep(1, ncol(Sigma)),
  ...
)
}
\arguments{
\item{Sigma}{empirical covariance matrix}

\item{B}{initial B matrix}

\item{C}{diagonal of intial C matrix}

\item{C0}{diagonal of penalization matrix}

\item{loss}{one of "loglik" (default) or "frobenius"}

\item{eps}{convergence threshold}

\item{alpha}{parameter line search}

\item{maxIter}{maximum number of iterations}

\item{lambda}{penalization coefficient for B}

\item{lambdac}{penalization coefficient for C}

\item{job}{integer 0,1,10 or 11}

\item{lambdas}{sequence of lambda}

\item{...}{additional arguments passed to \code{gclm}}
}
\value{
a list with the result of the optimization
}
\description{
Penalized-likelihood estimation of CLGGM
}
