% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clggm.R
\name{graddsllc}
\alias{graddsllc}
\title{Penalized likelihood estimation of CLGGM}
\usage{
graddsllc(Sigma, B, C = diag(ncol(Sigma)), C0 = diag(ncol(Sigma)),
  eps = 0.01, alpha = 0.5, beta = 0.2, maxIter = 1000,
  lambda = 0, job = 0)
}
\description{
Optimize the C matrix of a continuous Lyapunov
Gaussian graphical model (CLGGM) usign gradient descent.
\deqn{\hat{C} = \arg \min_C LL(B,C) + \lambda||C - C_0||_2^2}
}
