% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clggm.R
\name{graddsllc}
\alias{graddsllc}
\title{Penalized likelihood estimation of CLGGM}
\usage{
graddsllc(
  Sigma,
  B,
  C = diag(ncol(Sigma)),
  C0 = diag(ncol(Sigma)),
  eps = 0.01,
  alpha = 0.5,
  beta = 0.2,
  maxIter = 1000,
  lambda = 0,
  job = 0
)
}
\arguments{
\item{Sigma}{the observed covariance matrix}

\item{B}{an initial B matrix}

\item{C}{the C matrix}

\item{C0}{the C_0 matrix}

\item{eps}{convergence threshold for the proximal gradient}

\item{alpha}{backtracking parameter}

\item{beta}{backtracking parameter}

\item{maxIter}{the maximum number of iterations}

\item{lambda}{penalization coefficient}

\item{job}{integer}
}
\value{
a list with the output of the optimization:
\itemize{
\item \code{N}
\item \code{Sigma} the covariance of the estimated CLGGM
\item \code{B}
\item \code{C} the estimated C matrix
\item \code{C0}
\item \code{lambda}
\item \code{diff} the value of the last relative decrease
\item \code{objective} the value of the objective function
\item \code{iter} number of iterations
\item \code{job}
}
}
\description{
Optimize the diagonal C matrix of a continuous Lyapunov
Gaussian graphical model (CLGGM) usign gradient descent.
\deqn{\hat{C} = \arg \min_C LL(B,C) + \lambda||C - C_0||_2^2}
Subject to C diagonal positive definite.
}
