% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{llBpath}
\alias{llBpath}
\title{Path of B estimates}
\usage{
llBpath(Sigma, lambdas = NULL, C = diag(nrow(Sigma)), eps = 1e-08,
  maxIter = 1000, job = 10)
}
\arguments{
\item{Sigma}{empirical covariance matrix}

\item{lambdas}{increasing sequence of lambdas}

\item{C}{the known C =DD'}

\item{eps}{convergence criteria}

\item{maxIter}{maximum iterations for each proximal gradient}

\item{job}{integer flag (0,1,10,11)}
}
\description{
Path of B estimates
}
