% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lowertriangB}
\alias{lowertriangB}
\title{Recover lower triangular B}
\usage{
lowertriangB(Sigma, P = solve(Sigma), C = diag(nrow = nrow(Sigma)))
}
\arguments{
\item{Sigma}{PSD matrix, the covariance matrix}

\item{P}{PSD matrix, the inverse of the covariance}

\item{C}{PSD matrix}
}
\value{
A stable lower triangular matrix
}
\description{
Recover the only lower triangular stable matrix B such that the
associate OU process admit invariant distribution with covariance equal
to the given matrix.
}
